tts/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
boto3
pydub
speechkit
